WDFO-81 - create utils module
