+++ figo/figo.py
@@ -27,6 +27,7 @@ from .models import (
     User,
     WebhookNotification,
 )
+from .utils import filter_keys, filter_none, get_account_id
 from .version import __version__
 
 load_dotenv()
@@ -72,27 +73,6 @@ ERROR_MESSAGES = {
         "code": 90000,
     },
 }
-
-
-def get_account_id(account_or_account_id):
-    if account_or_account_id is None:
-        return None
-    elif isinstance(account_or_account_id, Account):
-        return account_or_account_id.account_id
-    else:
-        return account_or_account_id
-
-
-def filter_keys(options, allowed_keys):
-    if options is None or options == {}:
-        return {}
-    else:
-        keys = [key for key in options.keys() if key in allowed_keys]
-        return dict(zip(keys, [options[key] for key in keys]))
-
-
-def filter_none(dict_obj):
-    return {k: v for k, v in dict_obj.items() if v is not None}
 
 
 class FigoObject:
